introduce footer to typescript
