+++ src/components/Footers/FooterSmall.tsx
@@ -1,18 +1,26 @@
-import React from "react";
+interface AppProps {
+  absolute: string
+}
 
-export default function FooterAdmin() {
+export default function FooterSmall(props: AppProps) {
   return (
     <>
-      <footer className="block py-4">
+      <footer
+        className={
+          (props.absolute
+            ? "absolute w-full bottom-0 bg-blueGray-800"
+            : "relative") + " pb-6"
+        }
+      >
         <div className="container mx-auto px-4">
-          <hr className="mb-4 border-b-1 border-blueGray-200" />
+          <hr className="mb-6 border-b-1 border-blueGray-600" />
           <div className="flex flex-wrap items-center md:justify-between justify-center">
             <div className="w-full md:w-4/12 px-4">
               <div className="text-sm text-blueGray-500 font-semibold py-1 text-center md:text-left">
                 Copyright © {new Date().getFullYear()}{" "}
                 <a
-                  href="https://www.creative-tim.com?ref=nr-footer-admin"
-                  className="text-blueGray-500 hover:text-blueGray-700 text-sm font-semibold py-1"
+                  href="https://www.creative-tim.com?ref=nr-footer-small"
+                  className="text-white hover:text-blueGray-300 text-sm font-semibold py-1"
                 >
                   Creative Tim
                 </a>
@@ -22,32 +30,32 @@ export default function FooterAdmin() {
               <ul className="flex flex-wrap list-none md:justify-end  justify-center">
                 <li>
                   <a
-                    href="https://www.creative-tim.com?ref=nr-footer-admin"
-                    className="text-blueGray-600 hover:text-blueGray-800 text-sm font-semibold block py-1 px-3"
+                    href="https://www.creative-tim.com?ref=nr-footer-small"
+                    className="text-white hover:text-blueGray-300 text-sm font-semibold block py-1 px-3"
                   >
                     Creative Tim
                   </a>
                 </li>
                 <li>
                   <a
-                    href="https://www.creative-tim.com/presentation?ref=nr-footer-admin"
-                    className="text-blueGray-600 hover:text-blueGray-800 text-sm font-semibold block py-1 px-3"
+                    href="https://www.creative-tim.com/presentation?ref=nr-footer-small"
+                    className="text-white hover:text-blueGray-300 text-sm font-semibold block py-1 px-3"
                   >
                     About Us
                   </a>
                 </li>
                 <li>
                   <a
-                    href="http://blog.creative-tim.com?ref=nr-footer-admin"
-                    className="text-blueGray-600 hover:text-blueGray-800 text-sm font-semibold block py-1 px-3"
+                    href="http://blog.creative-tim.com?ref=nr-footer-small"
+                    className="text-white hover:text-blueGray-300 text-sm font-semibold block py-1 px-3"
                   >
                     Blog
                   </a>
                 </li>
                 <li>
                   <a
-                    href="https://github.com/creativetimofficial/notus-react/blob/main/LICENSE.md?ref=nr-footer-admin"
-                    className="text-blueGray-600 hover:text-blueGray-800 text-sm font-semibold block py-1 px-3"
+                    href="https://github.com/creativetimofficial/notus-react/blob/main/LICENSE.md?ref=nr-footer-small"
+                    className="text-white hover:text-blueGray-300 text-sm font-semibold block py-1 px-3"
                   >
                     MIT License
                   </a>
